Move profile sections to their own connected components
