+++ src/Containers/Favorites/Favorites.test.jsx
@@ -1,32 +1,33 @@
 import React from 'react';
+import { shallow } from 'enzyme';
 import TestUtils from 'react-dom/test-utils';
 import { Provider } from 'react-redux';
 import { MemoryRouter } from 'react-router-dom';
 import configureStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
-import Profile from './Profile';
+import sinon from 'sinon';
+import FavoritesContainer from './Favorites';
 
 const middlewares = [thunk];
 const mockStore = configureStore(middlewares);
 
-describe('Profile', () => {
+describe('FavoritesContainer', () => {
   it('is defined', () => {
     const profile = TestUtils.renderIntoDocument(<Provider store={mockStore({})}><MemoryRouter>
-      <Profile
-        isAuthorized={() => true}
-        onNavigateTo={() => {}}
-      />
+      <FavoritesContainer />
     </MemoryRouter></Provider>);
     expect(profile).toBeDefined();
   });
 
-  it('can handle authentication redirects', () => {
-    const profile = TestUtils.renderIntoDocument(<Provider store={mockStore({})}><MemoryRouter>
-      <Profile
-        isAuthorized={() => false}
-        onNavigateTo={() => {}}
-      />
-    </MemoryRouter></Provider>);
-    expect(profile).toBeDefined();
+  it('can call the onToggleFavorite function', () => {
+    const spy = sinon.spy();
+    const wrapper = shallow(
+      <FavoritesContainer.WrappedComponent
+        fetchData={() => {}}
+        toggleFavorite={spy}
+      />,
+    );
+    wrapper.instance().onToggleFavorite();
+    sinon.assert.calledOnce(spy);
   });
 });
